# backend/server/http/docs/swagger.yaml
definitions:
  auth.JwtToken:
    properties:
      expiration:
        type: integer
      grant_type:
        type: string
      token:
        type: string
    type: object
  auth.LoginPayload:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  auth.RefreshTokenPayload:
    properties:
      refresh_token:
        type: string
    type: object
  cashflow.CashFlow:
    properties:
      balance:
        type: number
      details:
        items:
          $ref: '#/definitions/cashflow.Detail'
        type: array
      total_payments:
        type: number
      total_sales:
        type: number
    type: object
  cashflow.Detail:
    properties:
      date:
        type: string
      description:
        type: string
      type:
        type: string
      value:
        type: number
    type: object
  dto.CreatePaymentPayload:
    properties:
      description:
        type: string
      value:
        type: number
    type: object
  dto.CreateProductPayload:
    properties:
      atacado_amount:
        type: integer
      name:
        type: string
      price_atacado:
        type: number
      price_varejo:
        type: number
    type: object
  dto.CreateUserPayload:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        type: string
    type: object
  dto.MessageJSON:
    properties:
      message:
        type: string
    type: object
  dto.RegisterSalePayload:
    properties:
      description:
        type: string
      items:
        items:
          $ref: '#/definitions/sales.CartItem'
        type: array
      payment_type:
        type: string
    type: object
  healthcheck.HealthStatus:
    properties:
      cache:
        type: string
      database:
        type: string
      status:
        type: string
      version:
        type: string
    type: object
  payment.Payment:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      value:
        type: number
    type: object
  product.Product:
    properties:
      atacado_amount:
        minimum: 1
        type: integer
      id:
        type: string
      name:
        type: string
      price_atacado:
        type: number
      price_varejo:
        type: number
    required:
    - name
    - price_atacado
    - price_varejo
    type: object
  sales.CartItem:
    properties:
      amount:
        type: integer
      item_id:
        type: string
    type: object
  sales.Item:
    properties:
      amount:
        type: integer
      name:
        type: string
      unit_price:
        type: number
    type: object
  sales.Sale:
    properties:
      date:
        type: string
      description:
        type: string
      id:
        type: string
      items:
        items:
          $ref: '#/definitions/sales.Item'
        type: array
      payment_type:
        type: string
      total:
        type: number
    type: object
  user.User:
    properties:
      email:
        type: string
      id:
        type: string
      name:
        type: string
    type: object
info:
  contact: {}
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Make login and get access-token
      parameters:
      - description: user info
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/auth.LoginPayload'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auth.JwtToken'
        "401":
          description: when email or password is invalid
          schema:
            $ref: '#/definitions/dto.MessageJSON'
        "422":
          description: when payload is invalid
          schema:
            $ref: '#/definitions/dto.MessageJSON'
        "500":
          description: when an error occurs
          schema:
            $ref: '#/definitions/dto.MessageJSON'
      summary: Login
      tags:
      - Auth
  /auth/me:
    get:
      consumes:
      - application/json
      description: Get current user data
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user.User'
        "500":
          description: when an error occurs
          schema:
            $ref: '#/definitions/dto.MessageJSON'
      summary: Get Me
      tags:
      - User
  /auth/refresh-token:
    post:
      consumes:
      - application/json
      description: Get a new access-token, this action will be expire the last one
      parameters:
      - description: the refresh token
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/auth.RefreshTokenPayload'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auth.JwtToken'
        "401":
          description: when token is invalid
          schema:
            $ref: '#/definitions/dto.MessageJSON'
        "500":
          description: when an error occurs
          schema:
            $ref: '#/definitions/dto.MessageJSON'
      summary: Refresh access-token
      tags:
      - Auth
  /cashflow:
    get:
      consumes:
      - application/json
      description: get cash flow
      parameters:
      - description: name search by q
        format: dateTime
        in: query
        name: start_at
        type: string
      - description: name search by q
        format: dateTime
        in: query
        name: end_at
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/cashflow.CashFlow'
        "400":
          description: when query is invalid
          schema:
            $ref: '#/definitions/dto.MessageJSON'
        "500":
          description: when an error occurs
          schema:
            $ref: '#/definitions/dto.MessageJSON'
      summary: Get cash flow
      tags:
      - Cashflow
  /health:
    get:
      description: Check app and dependencies status
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/healthcheck.HealthStatus'
      summary: Health Cehck
      tags:
      - Health check
  /payments:
    get:
      consumes:
      - application/json
      description: get payments
      parameters:
      - description: name search by q
        format: dateTime
        in: query
        name: start_at
        type: string
      - description: name search by q
        format: dateTime
        in: query
        name: end_at
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/payment.Payment'
            type: array
        "500":
          description: when an error occurs
          schema:
            $ref: '#/definitions/dto.MessageJSON'
      summary: List payments
      tags:
      - Payment
    post:
      consumes:
      - application/json
      description: Create a new payment and return payment data
      parameters:
      - description: the payload data
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/dto.CreatePaymentPayload'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/payment.Payment'
        "422":
          description: when payload is invalid
          schema:
            $ref: '#/definitions/dto.MessageJSON'
        "500":
          description: when an error occurs
          schema:
            $ref: '#/definitions/dto.MessageJSON'
      summary: Register a new payment
      tags:
      - Payment
  /payments/{id}:
    delete:
      consumes:
      - application/json
      description: Delete payment
      parameters:
      - description: the id of payment
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
        "500":
          description: when an error occurs
          schema:
            $ref: '#/definitions/dto.MessageJSON'
      summary: Delete Payment by ID
      tags:
      - Payment
    post:
      consumes:
      - application/json
      description: Update payment data
      parameters:
      - description: the id of payment
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/payment.Payment'
        "500":
          description: when an error occurs
          schema:
            $ref: '#/definitions/dto.MessageJSON'
      summary: Update Payment by ID
      tags:
      - Payment
  /products/:
    get:
      consumes:
      - application/json
      description: Get all products data
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/product.Product'
            type: array
        "500":
          description: when an error occurs
          schema:
            $ref: '#/definitions/dto.MessageJSON'
      summary: Get all products
      tags:
      - Product
    post:
      consumes:
      - application/json
      description: create a new product and return data
      parameters:
      - description: the product data
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/dto.CreateProductPayload'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/product.Product'
        "422":
          description: when data is invalid
          schema:
            $ref: '#/definitions/dto.MessageJSON'
        "500":
          description: when an error occurs
          schema:
            $ref: '#/definitions/dto.MessageJSON'
      summary: Create a New Product
      tags:
      - Product
  /products/{id}:
    delete:
      consumes:
      - application/json
      description: Delete product
      parameters:
      - description: the id of product
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
        "500":
          description: when an error occurs
          schema:
            $ref: '#/definitions/dto.MessageJSON'
      summary: Delete Product by ID
      tags:
      - Product
    get:
      consumes:
      - application/json
      description: Get product Data
      parameters:
      - description: the id of product
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/product.Product'
        "422":
          description: when id is invalid
          schema:
            $ref: '#/definitions/dto.MessageJSON'
        "500":
          description: when an error occurs
          schema:
            $ref: '#/definitions/dto.MessageJSON'
      summary: Get Product by ID
      tags:
      - Product
  /sales:
    get:
      consumes:
      - application/json
      description: get sales
      parameters:
      - description: name search by q
        format: dateTime
        in: query
        name: start_at
        type: string
      - description: name search by q
        format: dateTime
        in: query
        name: end_at
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/sales.Sale'
            type: array
        "400":
          description: when start or end param is invalid
          schema:
            $ref: '#/definitions/dto.MessageJSON'
        "500":
          description: when an error occurs
          schema:
            $ref: '#/definitions/dto.MessageJSON'
      summary: List sales
      tags:
      - Sale
    post:
      consumes:
      - application/json
      description: Register a sale and return sale data
      parameters:
      - description: the payload data
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/dto.RegisterSalePayload'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/sales.Sale'
            type: array
        "422":
          description: when payload is invalid
          schema:
            $ref: '#/definitions/dto.MessageJSON'
        "500":
          description: when an error occurs
          schema:
            $ref: '#/definitions/dto.MessageJSON'
      summary: Register a new sale
      tags:
      - Sale
  /sales/{id}:
    delete:
      consumes:
      - application/json
      description: Delete sale
      parameters:
      - description: the id of sale
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
        "500":
          description: when an error occurs
          schema:
            $ref: '#/definitions/dto.MessageJSON'
      summary: Delete Sale by ID
      tags:
      - Sale
  /users:
    post:
      consumes:
      - application/json
      description: Create a new user and return user data
      parameters:
      - description: the user data
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/dto.CreateUserPayload'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/user.User'
        "422":
          description: when payload is invalid
          schema:
            $ref: '#/definitions/dto.MessageJSON'
        "500":
          description: when an error occurs
          schema:
            $ref: '#/definitions/dto.MessageJSON'
      summary: Create User
      tags:
      - User
swagger: "2.0"